+++ lua/obsidian/lsp/handlers/did_rename_files.lua
@@ -1,7 +1,6 @@
 local obsidian = require "obsidian"
 local Note = obsidian.Note
 local api = obsidian.api
-local rename = require "obsidian.lsp.handlers._rename"
 
 ---@param file lsp.FileRename
 ---@param dispatchers table
@@ -9,12 +8,15 @@ local function rename_note(file, dispatchers)
   local new_path = vim.uri_to_fname(file.newUri)
   local note = Note.from_file(new_path)
   local new_name = vim.fs.basename(new_path):gsub("%.md$", "")
-  rename.build_edit(note, new_name, {
+  note:rename(new_name, {
     old_path = vim.uri_to_fname(file.oldUri),
     new_path = new_path,
     include_file_rename = false,
-  }, function(edit, meta)
-    if not edit then
+    apply = false,
+    update_buffers = false,
+    check_unique = false,
+  }, function(err, edit, meta)
+    if err or not edit then
       return
     end