+++ lua/obsidian/lsp/handlers/rename.lua
@@ -5,8 +5,6 @@ local log = obsidian.log
 local api = obsidian.api
 local util = obsidian.util
 
-local M = require "obsidian.lsp.handlers._rename"
-
 ---@param params lsp.RenameParams
 return function(params, handler, _)
   local new_name = params.newName
@@ -20,16 +18,12 @@ return function(params, handler, _)
   local cur_link = api.cursor_link()
 
   local function do_rename(note)
-    local old_stem = note.path and note.path.stem or nil
-    if new_name == note.id or (old_stem and new_name == old_stem) then
-      log.info "Identical name"
-      return handler(nil, {})
-    end
-    if not M.validate(new_name) then
-      log.info "Note with same name exists"
-      return handler(nil, {})
-    end
-    M.rename(note, new_name, handler)
+    note:rename(new_name, { apply = false }, function(err, edit)
+      if err or not edit then
+        return handler(nil, {})
+      end
+      handler(nil, edit)
+    end)
   end
 
   if cur_link then